Break the nested-ragged chain with a `RaggedFromNested` base.
`RaggedFromNestedRowSplits extends RaggedFromNestedRowLengths` to borrow
its nested-ragged shape logic, overriding only the partitioning-argument
source and the row-dimension computation — a code-reuse-not-is-a chain
(wala/ML#514). `from_nested_row_splits` is not a kind of
`from_nested_row_lengths`; they are siblings.

Extract the shared shape construction into an abstract `RaggedFromNested`
base (flat_values at position 0, the nested partitioning argument at
position 1) with abstract `getNestedStructureParameterName()` and
`computeRowDim()`. Both concrete forms extend it: row-lengths uses the
first-element length directly; row-splits subtracts one. Behavior-
preserving (positions and names unchanged); full suite green (970).
