+++ com.ibm.wala.cast.python.ml/source/com/ibm/wala/cast/python/ml/client/RaggedFromNestedRowSplits.java
@@ -2,52 +2,26 @@ package com.ibm.wala.cast.python.ml.client;
 
 import com.ibm.wala.cast.python.ml.types.TensorType.Dimension;
 import com.ibm.wala.cast.python.ml.types.TensorType.NumericDim;
-import com.ibm.wala.ipa.callgraph.propagation.InstanceKey;
 import com.ibm.wala.ipa.callgraph.propagation.PointsToSetVariable;
-import com.ibm.wala.ipa.callgraph.propagation.PropagationCallGraphBuilder;
-import com.ibm.wala.util.intset.OrdinalSet;
-import java.util.Locale;
-import java.util.logging.Logger;
 
 /**
- * A representation of the `tf.RaggedTensor.from_nested_row_splits` API in TensorFlow.
+ * A representation of the `tf.RaggedTensor.from_nested_row_splits` API in TensorFlow. The
+ * nested-ragged shape construction lives in {@link RaggedFromNested}; this form reads the {@code
+ * nested_row_splits} argument and subtracts one from its first-element length ({@code row_splits}
+ * has {@code nrows + 1} entries).
  *
  * @see <a
  *     href="https://www.tensorflow.org/api_docs/python/tf/RaggedTensor#from_nested_row_splits">tf.RaggedTensor.from_nested_row_splits</a>.
  */
-public class RaggedFromNestedRowSplits extends RaggedFromNestedRowLengths {
-
-  @SuppressWarnings("unused")
-  private static final Logger LOGGER = Logger.getLogger(RaggedFromNestedRowSplits.class.getName());
-
-  protected enum Parameters {
-    FLAT_VALUES,
-    NESTED_ROW_SPLITS,
-    NAME,
-    VALIDATE;
-
-    public String getName() {
-      return name().toLowerCase(Locale.ROOT);
-    }
-
-    public int getIndex() {
-      return ordinal();
-    }
-  }
+public class RaggedFromNestedRowSplits extends RaggedFromNested {
 
   public RaggedFromNestedRowSplits(PointsToSetVariable source) {
     super(source);
   }
 
-  protected String getNestedRowSplitsParameterName() {
-    return Parameters.NESTED_ROW_SPLITS.getName();
-  }
-
   @Override
-  protected OrdinalSet<InstanceKey> getNestedStructurePointsToSet(
-      PropagationCallGraphBuilder builder) {
-    return this.getArgumentPointsToSet(
-        builder, this.getNestedRowLengthsParameterPosition(), getNestedRowSplitsParameterName());
+  protected String getNestedStructureParameterName() {
+    return "nested_row_splits";
   }
 
   @Override
